modifying the classes
