+++ build.gradle
@@ -8,6 +8,7 @@ repositories {
     // Use jcenter for resolving your dependencies.
     // You can declare any Maven/Ivy/file repository here.
     jcenter()
+    mavenCentral()
 }
 
 
@@ -41,16 +42,9 @@ test {
 dependencies {
     compile 'org.codehaus.groovy:groovy-all:2.5.2'
     compile 'com.google.inject:guice:4.1.0'
-
     compile group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '3.141.59'
-
     compile group: 'org.awaitility', name: 'awaitility', version: '3.1.2'
     compile group: 'org.seleniumhq.selenium', name: 'htmlunit-driver', version: '2.27'
-
-// https://mvnrepository.com/artifact/org.junit.jupiter/junit-jupiter-api
- //   compile group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.4.0'
     testImplementation("org.junit.jupiter:junit-jupiter-api:5.4.2")
     testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.4.2")
- //   compile group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.4.0'
-
 }
