+++ src/main/java/dotdash/automation/pageobjects/NewWindowPageObject.java
@@ -1,16 +1,21 @@
 package dotdash.automation.pageobjects;
 
 import dotdash.automation.ui.BasePageObject;
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.PageFactory;
 
 public class NewWindowPageObject extends BasePageObject {
-    private By text = By.cssSelector("div h3");
+    @FindBy(css = "div h3")
+    private WebElement text;
+
     public NewWindowPageObject(WebDriver driver) {
         super(driver);
+        PageFactory.initElements(getDriver(), this);
     }
 
-    public String getText(){
-        return getDriver().findElement(text).getText();
+    public String getText() {
+        return text.getText();
     }
 }
