+++ src/main/java/dotdash/automation/pageobjects/NotificationMessagePageObject.java
@@ -17,7 +17,7 @@ public class NotificationMessagePageObject extends BasePageObject {
 
     public NotificationMessagePageObject(WebDriver driver) {
         super(driver);
-        PageFactory.initElements(driver,this);
+        PageFactory.initElements(getDriver(),this);
     }
 
     public String getMessage(){
