+++ src/main/java/dotdash/automation/pageobjects/WindowsPageObject.java
@@ -3,20 +3,23 @@ package dotdash.automation.pageobjects;
 import dotdash.automation.ui.BasePageObject;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
 
 import java.util.ArrayList;
 
 public class WindowsPageObject extends BasePageObject {
-    private By clickHere = By.linkText("Click Here");
+    @FindBy(linkText = "Click Here")
+    private WebElement clickHere;
 
     public WindowsPageObject(WebDriver driver) {
         super(driver);
     }
 
     public NewWindowPageObject clickHere() {
-        getDriver().findElement(clickHere).click();
-        ArrayList<String> tabs2 = new ArrayList<>(getDriver().getWindowHandles());
-        driver.switchTo().window(tabs2.get(1));
+        clickHere.click();
+        ArrayList<String> tabs = new ArrayList<>(getDriver().getWindowHandles());
+        driver.switchTo().window(tabs.get(1));
         return new NewWindowPageObject(driver);
     }
