+++ src/main/java/dotdash/automation/ui/BasePageObject.java
@@ -1,12 +1,7 @@
 package dotdash.automation.ui;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
-
-import java.util.concurrent.TimeUnit;
 
 public class BasePageObject {
 
@@ -22,6 +17,7 @@ public class BasePageObject {
     public WebDriver getDriver() {
         return driver;
     }
+
     public String getTitle() {
         return driver.getTitle();
     }
